ingredientes/api/serializers.py
from rest_framework import serializers
from ingredientes.models import Ingrediente

class IngredienteSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Ingrediente
        fields = ['id', 'nombre', 'descripcion',]